Add assignment typecasts

+++ iothub_client/src/iothubtransport_mqtt_common.c
@@ -480,7 +480,7 @@ static int parse_device_twin_topic_info(const char* resp_topic, bool* patch_msg,
                 }
                 else if (token_count == 3)
                 {
-                    *status_code = atol(STRING_c_str(token_value));
+                    *status_code = (int)atol(STRING_c_str(token_value));
                     if (STRING_TOKENIZER_get_next_token(token_handle, token_value, "/?$rid=") == 0)
                     {
                         *request_id = (size_t)atol(STRING_c_str(token_value));
